v0.2.2 Layout, types, formatting
# Formatted to 3 space indents and set route Layout

Modified the prettier config to use 3 spaces for tab indents. This is to
give enough space to identify adjacent react elements.

Refactored the .vscode settings to include the new tab size and more
comments.

Created the Layout component to be used as a parent route for global
navbar and footer for all child routes.

Refactored some naming conventions for pages.

Extracted types from files and moved them into the types module.

Added a backdrop filter blur on the navbar.

# Refactored Products, types, and components

Refactored the Products page to include products.

Refactored the ProductCard into two seperate cards, FeaturedProductCard
and ProductCard.
- Tweaked the styles a bit

Refactored the types so it is a module.

Added new sample data to test new design for Products page.