v0.4.0 Revisions, CartDrawer, Preloader, formatting
Formatted files
- Combined all vendor imports together and only seperate vendor
modules and the ones I created.

Added custom scrollbar, refactored index.css

Added simplebar-react module

Created the Scrollbar component which is an higher order component of
simplebar-react to pass in the theme mode to change the color of the
bar.

Removed the properties on the body element since Mui takes care of that
with theming.

v0.3.1 State, CartDrawer, Nav

Created a code snippet file
- Imports redux hooks
- Shorthand, no fluff, Mui Typography component

Moved the types for Navbar and others into their files and export them.

Moved the NavbarItemProps type into the routes.ts and exported.

Moved the NavButton component inside of the Nav module.

Overrode Mui default color on buttons to black for contrasting.

Added the CartDrawer to the Layout component.

Improved index.css
- Removed decoration on all anchor elements

Refactored all uses of the "white" color to "text.primary" for future
compatibility with light/dark modes.

Added a Preloader

Added zoom effect on item thumbnails
- Gives the user a hint that they can click on the thumbnails, it zooms
on hover.

Fixed start script
- The semicolon on line 127 was needed.