v0.5.0 Full state implementation and cart features
Removed the lint:css script since stylelint and such were removed.

Removed the run for pull requests on the main branch since linting
isn't too important on pulls.

Added Toast component

Fixed new test directory for jest config
- The directory became __tests__ from test and requires a config setting.

Refactored product routes

Using the product name for the url instead of the id so that the user
will better understand the page they're on.

Using an extraReducer to save the cart as json to localStorage on any
change.

Compeleted redux state management for the cart drawer and to retrieve
data and images from the server for the products.

Added axios for handling requests

Updated eslintrc
- The space-before-function-paren arrow an anonymous functions require a
space while named functions don't

Renamed tests folder to __tests__ so those files don't affect the
webpack bundling.

Added proxy for backend in package.json

Removed Stylelint
- All the styles are themed through Mui or through the sx prop for on-the
-fly styling on components so there aren't many stylesheets to lint.

Set the TextField on the Quote component of the Home page to fullWidth
since it is contained in a grid item.

Changed the Add To Cart button to inherit the light/dark color.

Added more spacing between the description and Add To Cart button as
well as a little header.

Update package.json version