v0.5.1 404 page and fixes
Fixed analyze script
- Was missing pushing the the webpack bundle analyzer in the array
of plugins.

Installed dev dep progress-bar-webpack-plugin to view progress.
- Added progress bar to build script

Installed gsap for 404 animation

Update CartDrawer handlers
- Added the handleClose function when clicking the cart buttons
that redirect to another page.

Added a 404 page.

Added the Cart page which is in process of being completed.

Update theme to set mode

Update CartItems to use name for item link
- Was originally using the id of the product but now using the name
for better urls to undertand the page the user is on.

Update tests
- Added a simple test that will pass in the setupTests file to stop
the error occuring saying that the file requires a test; this is
just importing the library.