v0.6.0 Update API requests
Created /api directory to manage requests with axios. Because the
backend is hosted on a different domain (Heroku) we need to specify the
URL and concanetate the endpoints to get the true request URL.

Added the feature property for products for the featured items on home
page.

Fixed home page input fields autocomplete

For some reason the box-shadow was set to blue on them. Set to none in
index.css