v0.9.0 Update logos and CI
Removed the oldNavbar and logo images

Removed the eslintconfig that was migrated from the package.json from
CRA since it provides no rules that help.

Created and set new svg logos that will be pulled from the static assets
since they are scalable.

Refactored navbar
- Using a constant for the logo from static assets
- Set the toolbar to a height of 80px
- Added aria-label to nav buttons using the route name

Cannot use the "homepage" property in the package.json because it
conflicts with the webpack build process.

Set fsevents as optionalDep to fix the Github actions need to move
it because it causes issues.

Update CI Github actions
- Removed support for the 12.x node version since netlify uses 14.x.