v0.12.0 Fixed husky for pre-commit linting
Wasn't setup for this project so I looked into it and saw that it pretty
much had to be installed again.

Added the eslintcache to the .gitignore so it isn't added to the
repository since it's just used to only lint files that have changed
since the last lint was run for performance.

Now, it will prevent commits for silly or tiny errors in the code. This
will greatly improve the development process since it won't waste time
with correctable errors in the Github action CI/CD pipeline.