v0.13.0 Mobile ready, preview deploys, routing
Refactored ProductCard
- Set the styledCard component to width of 100% to fill spacing of
container
- Modified text sizes to be smaller for the product name

Refactored 404 page for mobile

Created a deploy script in the package.json to be used to deploy a
preview online.

Updated number in footer

Refactored the Navbar for the mobile and the mobile menu is now included
with it for the menu open slide down effect, and it's small.

Removed the lint job from the Gitbub action CI since the husky and
lint-staged combo will prevent commits by checking it. Plus, it caches
files so it improves performance by only linting files that changed.

Modified the primary color for the theme.

Refactored Home page to support mobile viewing.

Updated the routes file to properly reflect the routes available
and to be mapped for nav routes.

Refactored App.tsx to map the routes into Route elements.