v0.14.0 Fixed logo issues
The logos were moved from the public directory to the src/assests/img
where they can be imported for the navbars.

Fixed the logo svg files that included an AdobeIllustrator namespace tag
on the g elements, causing an error to be thrown.

Used SVGminify.com to minify the svg files to save even more space.