v0.18.0 Refactored design, modified deploy script
Refactored design, improvement

Refactored Gallery component for Item page:
- Wrapped the thumbnails for the gallery in the simple-react-lightbox
wrapper and the entire component is contained within the context
provider.
- Removed the Grid and used flexbox with Box components

Refactored index.tsx
- Removed the simple-react-lightbox module wrapping the entire app to
just the Gallery component where it is used.

Modified mobile design on 404 and products pages.

Fixed Navbar mobilemenu dropdown:
- The menu slide needs to stay mounted
- The list item navigation unmountOnExit to slide in properly each time
menu is opened

Refactored deploy script: was performing a delete before building. It is
a test deployment so it just runs a prod build and deploys it.