v0.26.0 Update workflows; dependabot automerge on PR
Added auto-merge.yml to run on pull requests on the main branch. First
checks if the actor is the dependabot to stop the action if it isn't. It
will then fetch the metadata after waiting for the result of the other
actions (build and test). Only if the results of the those actions are
successful will it then proceed to approve the PR and then perform a
squash merge.

The publish workflow is disabled because publishing directly is easier.
A new branch, CI/CD will be used to automatically run the actions
(build, lint, test). If those pass when pushed, then pushing the main
branch will trigger the publish.