v0.31.0 Refactor components for barrel exports
Instead of having multiple tabs with index.tsx, which can be
confusing, using barrel exports so that the index.js file does the
actual export while the actual component will be the name, i.e.,
CartDrawer.tsx.

Update naming on other components like spinners and 404 page to
PageNotFound for clarity.