v0.32.0 Updates and fixes
Set the URL constant with an envrionment variable for the API.
-- .env files updated with new variable
-- api/index.ts updated

By specifying the localhost with port 4000, the proxy in package.json is
no longer needed. This also prevents any errors when hot reloading, to
not have it force to the proxy.
-- package.json

analyze.js
- Minor fixes and additions
- Removed the newline character in the format property

Moved setupTests from __tests__ to the src directory and updated the
Jest settings in package.json to reflect this. This is so that when
running the tests, it will just setup the environment, not be run as a
test.
-- Removed the useless test to make it pass in previous configuration
-- Updated package.json Jest setting

Refactored app to modularize the reducers and middleware
-- reducers/cart and reducers/products slices
-- new middleware.ts
-- Updated imports for files

Refactor ProductCard
- Added a style prop to Typography component for the price so that the
width is properly adjusted for the skeleton.

Refactor Home/Courses section
- Updated the iFrame for the video to allow fullscreen