v0.33.0 Update workflows and testing
Updated the Node version to use the current version I'm developing with
of 17.x for the workflows.

Updated netlify.toml config file to specify the Node version as an
envrionment variable.

Updated Jest config to only test files with the spec or test
extensions to stop it from running tests on the mock file.
-- Moved the setupTests.js back into the __tests__ directory.