v0.44.0 Courses page and updates
Turned off dependabot by commenting out the code.

Created and added a BackToTop component to scroll to the top into
`App.tsx`.

Created the courses slice for state management, similar to the products.

Added the content to the courses page with basic placeholder cards
similar to the products.

Updated the types.

Added error messages for the home products section and products page for
when failing to fetch data.